requirements.txt
# Core
httpx[http2]>=0.28.1
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
tenacity>=9.0.0

# Database
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# UI/CLI
rich>=13.9.0
questionary>=2.1.0

# Future / Stage 2 & 3 (Optional for now)
# openai>=1.50.0  # Or google-generativeai
